feat(slack-app): typings for manifest
